refactor: move TokenToggleInfo into select-travel-token-screen
