+++ src/select-travel-token-screen/index.ts
@@ -1,1 +1,2 @@
 export {SelectTravelTokenScreenComponent} from './SelectTravelTokenScreenComponent';
+export {TokenToggleInfo} from './TokenToggleInfo';
