--- src/token-toggle-info/index.ts
@@ -1,1 +1,0 @@
-export {TokenToggleInfo} from './TokenToggleInfo';
